user-microservice: funcionalidad listar usuarios

+++ user-service/src/main/java/com/yeisonmenau/user_service/infrastructure/web/dto/response/UserResponseDTO.java
@@ -11,5 +11,5 @@ public class UserResponseDTO {
     private Long userId;
     private String username;
     private String email;
-    private Long edad;
+    private Integer edad;
 }
